+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Changes to prior versions can be found on the [Github release page](https://gith
 
 ### Changed
 * Require JDK 25
+* Pin GitHub action versions used in CI ([#132](https://github.com/cryptomator/integrations-linux/pull/132))
+* Updated dependency `com.fasterxml.jackson.core:jackson-databind` from 2.20.0 to 2.20.1
 
 ## [1.6.1](https://github.com/cryptomator/integrations-linux/releases/tag/1.6.1) - 2025-09-17